% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase.R
\name{erase}
\alias{erase}
\title{Erase \code{scribblr} notes}
\usage{
erase()
}
\description{
Deletes the \code{scribblr} note file associated to the active
RStudio project, if any, otherwise to the global RStudio session.
}
\examples{
\dontrun{
erase()
}
}
\seealso{
\link{scribble}
}
\author{
Valerio Gherardi
}
