% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudiomemo-package.R
\docType{package}
\name{rstudiomemo-package}
\alias{rstudiomemo}
\alias{rstudiomemo-package}
\title{rstudiomemo:}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
