% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudionotes-package.R
\docType{package}
\name{rstudionotes-package}
\alias{rstudionotes}
\alias{rstudionotes-package}
\title{rstudionotes:}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
