% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scribble.R
\name{scribble}
\alias{scribble}
\title{Open \code{scribblr} note editor}
\usage{
scribble(note = NULL)
}
\arguments{
\item{note}{either \code{NULL}, or a length one character (not \code{NA}).
See details.}
}
\value{
returns \code{NULL}, invisibly. Called for side-effects.
}
\description{
Opens the \code{scribblr} note editor in a new window.
}
\details{
\code{scribblr} integrates a minimalist note editor within RStudio,
useful for taking quick project-related notes without distractions.
\code{scribblr} notes are RStudio project aware: each project has associated
its own notes, which can be accessed by calling \code{scribble()}, and are
stored into the project's root directory. Using \code{scribble()} without
any active project will take the R home directory as root.

Calling \code{scribble()} with the default \code{note = NULL} gives access to
the main project notes. Otherwise, \code{note} must be a string specifying a
valid filename.

\code{scribblr} notes and settings for the active project are stored in the
\code{".scribblr"} directory, under the project's root. If this, or the note
specified by \code{note}, do not exist, the user will be prompted for
permission to create the required files/directories.

Notes are autosaved when the editor is closed; until that
moment, the R session will remain busy.

\code{scribble()} can also be called (and, in particular, associated a
custom keystroke) via the RStudio Addin "Open scribblr note editor".
}
\examples{
\dontrun{
scribble()
}
}
\author{
Valerio Gherardi
}
