% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scribblr-package.R
\docType{package}
\name{scribblr-package}
\alias{scribblr}
\alias{scribblr-package}
\title{scribblr: A Minimalist Notepad Inside RStudio}
\description{
a minimalist notepad for taking project-related notes in RStudio. Registers an RStudio Addin
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/scribblr}
  \item Report bugs at \url{https://github.com/vgherard/scribblr/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
