% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_management.R
\name{scribblr_delete}
\alias{scribblr_delete}
\title{Delete a \code{scribblr} note}
\usage{
scribblr_delete(note)
}
\arguments{
\item{note}{a length one character (not NA). Note to be deleted.}
}
\value{
returns \code{NULL}, invisibly. Used for side-effects.
}
\description{
Delete a \code{scribblr} note.
}
\examples{
\dontrun{
erase()
}
}
\seealso{
\link{scribble}
}
\author{
Valerio Gherardi
}
