% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_management.R
\name{scribblr_list}
\alias{scribblr_list}
\title{List \code{scribblr} notes}
\usage{
scribblr_list()
}
\value{
a character vector.
}
\description{
List all \code{scribblr} notes.
}
\examples{
\dontrun{
scribblr_list("todo")
}
}
\seealso{
\link{scribble}
}
\author{
Valerio Gherardi
}
