% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase_notes.R
\name{erase_notes}
\alias{erase_notes}
\title{Erase \code{scribblr} notes}
\usage{
erase_notes()
}
\description{
Deletes the \code{scribblr} note file associated to the active
RStudio project (or to the global RStudio session, if no project is active).
}
\examples{
\dontrun{
erase()
}
}
\seealso{
\link{scribble}
}
\author{
Valerio Gherardi
}
