% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scribblr-package.R
\docType{package}
\name{scribblr-package}
\alias{scribblr}
\alias{scribblr-package}
\title{scribblr: Notetaking for RStudio Projects}
\description{
RStudio Addin integrating a minimalist notepad in the RStudio IDE.
}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
