% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scribblr-package.R
\docType{package}
\name{scribblr-package}
\alias{scribblr}
\alias{scribblr-package}
\title{scribblr: A Notepad Inside RStudio}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A project aware notepad inside RStudio, for taking quick project-related notes without distractions. RStudio addin.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/vgherard/scribblr/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
